fix(app): add fix to issues noticed from API tests
